utf8n_to_uvchr() Fix EBCDIC bug with overlongs
The comment removed in this commit was wrong, and so was the code it
described.  On EBCDIC platforms, there are malformations that need to be
converted from Unicode to native.  When I wrote that I wasn't thinking
about overlongs, which can evaluate to any code point.  The new tests in
d566bd20c27a46aecd668d2f739b9515f46ac74f caught this.