the test for #129897 was missing the min-mod (?) on the dot star
A plain .* will make the pattern match skip the bug we
are testing for. We need .*? to see the difference

Fixed:
./perl -Ilib -le'"riiaan"=~/(.*?(a(a)|i(i))n)/ and print "$2-$3-$4-$1"'
aa-a--riiaan

Broken:
perl -le'"riiaan"=~/(.*?(a(a)|i(i))n)/ and print "$2-$3-$4-$1"'
aa-a-i-riiaan

Maybe the test could make this more explicit. Not sure if its
worth it.

Thanks to Aaron Crane for noticing the test didn't do what it was
meant to. Bad Yves.