Check for echo in new Net::Ping tests.
Net::Ping->new() with no parameters calls

  getservbyname('echo', 'tcp');

and falls down hard if that call does not succeed.  It would be
unusual to be running an echo service on any well-maintained
system on any platform.  On VMS at least, doing a name lookup on
a service that is not configured and enabled returns an error, and
thus any test that uses this ancient default will fail.  So skip
those tests if the getservbyname() fails.