Eliminate SVrepl_EVAL and SvEVALED()
This flag is only used to indicate that the SV holding the text of the
replacement part of a s/// has seen at least one /e.

Instead, set the IVX field in the SV to a true value.
(We already set the NVX field on that SV to indicate a multi-src-line
substitution).

This is to reduce the number of odd special cases for the SVpbm_VALID flag.