Add indented here-docs.
This adds a new modifier '~' to here-docs that tells the parser
that it should look for /^\s*$DELIM\n/ as the closing delimiter.

These syntaxes are all supported:

  <<~EOF;
  <<~\EOF;
  <<~'EOF';
  <<~"EOF";
  <<~`EOF`;
  <<~ 'EOF';
  <<~ "EOF";
  <<~ `EOF`;

The '~' modifier will strip, from each line in the here-doc, the
same whitespace that appears before the delimiter.

Newlines will be copied as is, and lines that don't include the
proper beginning whitespace will cause perl to croak.

Some examples:

  if (1) {
    print <<~EOF;
      Hello there
      EOF
  }

prints "Hello there\n";

The following:

  if (1) {
    print <<~EOF;
     Hello There
      EOF
  }

croaks with:

  Indentation on line 1 of here-doc doesn't match delimiter at - line 2.