make sure that Encode uses SvPV_force style macros when appropriate
This fixes two bugs. The first is in Encode.xs where we
would handle COW strings properly, but not "unstealable" strings
(where the SvLEN == 0).

The second is in Unicode.xs where we would not handle COW
nor "unstealable" strings.

Both functions should be using a "force" style SvPV macro when
they intend to modify the string in any way.

This needs to go upstream.