perlre: summarise full syntax for (?(cond)then|else) constructs
In the conditional-execution constructs, the condition is always syntactically
surrounded by a single pair of parens. The various constructs therefore show
that pair of parens in all cases; this seems like a good thing. In addition,
the summary for these constructs as a group also shows the parens; this also
seems like a good thing. But it's not immediately obvious that the two sets
of parens are the same.

Rather than trying to clarify the situation using complicated prose, just
show an example of the full syntax for each conditional construct.