perlre: don't impugn the stability of all (?…) constructs
Previously, the documentation suggested that any or all of these constructs
might disappear or be significantly changed without notice:

    (?#comment)
    (?m)
    (?^m)
    (?:group)
    (?m-s:group)
    (?^m:group)
    (?|branch|reset)
    (?=pos lookahead)
    (?!neg lookahead)
    (?<=pos lookbehind)
    (?<!neg lookbehind)
    (?<named>capture)
    (?'named'capture)
    \k<named_backref>
    \k'named_backref'
    (?{ code })
    (??{ postponed regex })
    (?1)
    (?-2)
    (?+2)
    (?R)
    (?0)
    (?&named_subpattern)
    (?(condition)consequent|alternate), including (?(DEFINE)defs)
    (?>independent subpattern)
    (?[ [ext] + [char] - [cls] ])

Of those features, the last one (extended bracketed character classes) is
specifically experimental in our formal sense; but it's not realistic to
think that future versions of Perl might break any of the others — many of
which date back as far as 5.000. Furthermore, even if that were likely
enough to be worth pointing out, it would be better to do so on each of the
affected constructs, rather than with an easier-to-miss blanket notice at
the top of the section.

Therefore, this change removes the blanket notice, and adds a note of
experimental status to the mention of extended bracketed character classes,
linking to our policy definition of what that means.